refactor(ingester-example): reorganize project structure and improve database connection
- Rename packages to better reflect their purpose (e.g., quickstart, benchmark)
- Move database connection logic to a central DBConnector class
- Use properties file for database connection details
- Update logging and error handling
- Improve code readability and maintainability
