+++ ingester-example/src/main/java/io/greptime/bench/benchmark/StreamingWriteBenchmark.java
@@ -14,30 +14,29 @@
  * limitations under the License.
  */
 
-package io.greptime.bench;
+package io.greptime.bench.benchmark;
 
 import io.greptime.GreptimeDB;
-import io.greptime.WriteOp;
+import io.greptime.StreamWriter;
+import io.greptime.bench.DBConnector;
+import io.greptime.bench.TableDataProvider;
 import io.greptime.common.util.MetricsUtil;
 import io.greptime.common.util.ServiceLoader;
 import io.greptime.common.util.SystemPropertyUtil;
 import io.greptime.metrics.ExporterOptions;
 import io.greptime.metrics.MetricsExporter;
-import io.greptime.models.Err;
-import io.greptime.models.Result;
 import io.greptime.models.Table;
 import io.greptime.models.TableSchema;
 import io.greptime.models.WriteOk;
 import io.greptime.rpc.Compression;
 import io.greptime.rpc.Context;
-import java.util.Arrays;
 import java.util.Iterator;
 import java.util.concurrent.CompletableFuture;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * BatchingWriteBenchmark is a benchmark for the batching write API of GreptimeDB.
+ * StreamingWriteBenchmark is a benchmark for the streaming write API of GreptimeDB.
  *
  * Env:
  * - db_endpoint: the endpoint of the GreptimeDB server
@@ -46,17 +45,14 @@ import org.slf4j.LoggerFactory;
  * - zstd_compression: whether to use zstd compression
  * - max_points_per_second: the max number of points that can be written per second, exceeding which may cause blockage
  */
-public class BatchingWriteBenchmark {
+public class StreamingWriteBenchmark {
 
-    private static final Logger LOG = LoggerFactory.getLogger(BatchingWriteBenchmark.class);
+    private static final Logger LOG = LoggerFactory.getLogger(StreamingWriteBenchmark.class);
 
     public static void main(String[] args) throws Exception {
-        String endpoint = SystemPropertyUtil.get("db_endpoint", "127.0.0.1:4001");
-        String dbName = SystemPropertyUtil.get("db_name", "public");
         boolean zstdCompression = SystemPropertyUtil.getBool("zstd_compression", true);
         int batchSize = SystemPropertyUtil.getInt("batch_size_per_request", 64 * 1024);
         int maxPointsPerSecond = SystemPropertyUtil.getInt("max_points_per_second", Integer.MAX_VALUE);
-        LOG.info("Connect to db: {}, endpoint: {}", dbName, endpoint);
         LOG.info("Using zstd compression: {}", zstdCompression);
         LOG.info("Batch size: {}", batchSize);
         LOG.info("Max points per second: {}", maxPointsPerSecond);
@@ -65,10 +61,12 @@ public class BatchingWriteBenchmark {
         MetricsExporter metricsExporter = new MetricsExporter(MetricsUtil.metricRegistry());
         metricsExporter.init(ExporterOptions.newDefault());
 
-        GreptimeDB greptimeDB = DBConnector.connectTo(new String[] {endpoint}, dbName);
+        GreptimeDB greptimeDB = DBConnector.connect();
 
         Compression compression = zstdCompression ? Compression.Zstd : Compression.None;
         Context ctx = Context.newDefault().withCompression(compression);
+
+        StreamWriter<Table, WriteOk> writer = greptimeDB.streamWriter(maxPointsPerSecond, ctx);
 
         TableDataProvider tableDataProvider =
                 ServiceLoader.load(TableDataProvider.class).first();
@@ -90,21 +88,21 @@ public class BatchingWriteBenchmark {
             LOG.info("Table bytes used: {}", table.bytesUsed());
             // Complete the table; adding rows is no longer permitted.
             table.complete();
-            long fStart = System.nanoTime();
             // Write the table data to the server
-            CompletableFuture<Result<WriteOk, Err>> future =
-                    greptimeDB.write(Arrays.asList(table), WriteOp.Insert, ctx);
-            // Wait for the write to complete
-            int numRows = future.get().mapOr(0, writeOk -> writeOk.getSuccess());
-            long costMs = (System.nanoTime() - fStart) / 1000000;
-            LOG.info("Write rows: {}, time cost: {}ms", numRows, costMs);
+            writer.write(table);
 
             if (!rows.hasNext()) {
                 break;
             }
         }
 
-        LOG.info("Completed writing data, time cost: {}s", (System.nanoTime() - start) / 1000000000);
+        // Completes the stream, and the stream will be closed.
+        CompletableFuture<WriteOk> future = writer.completed();
+
+        // Now we can get the write result.
+        WriteOk result = future.get();
+
+        LOG.info("Completed writing data: {}, time cost: {}s", result, (System.nanoTime() - start) / 1000000000);
 
         greptimeDB.shutdownGracefully();
         tableDataProvider.close();
