+++ ingester-example/src/main/java/io/greptime/quickstart/write/BulkWriteApiQuickStart.java
@@ -14,13 +14,16 @@
  * limitations under the License.
  */
 
-package io.greptime;
+package io.greptime.quickstart.write;
 
+import io.greptime.BulkStreamWriter;
 import io.greptime.BulkWrite.Config;
+import io.greptime.GreptimeDB;
 import io.greptime.common.util.StringBuilderHelper;
 import io.greptime.models.DataType;
 import io.greptime.models.Table;
 import io.greptime.models.TableSchema;
+import io.greptime.quickstart.TestConnector;
 import io.greptime.rpc.Compression;
 import io.greptime.rpc.Context;
 import java.math.BigDecimal;
