+++ ingester-example/src/main/java/io/greptime/quickstart/write/HighLevelApiStreamWriteQuickStart.java
@@ -14,9 +14,15 @@
  * limitations under the License.
  */
 
-package io.greptime;
+package io.greptime.quickstart.write;
 
+import io.greptime.metric.Cpu;
+import io.greptime.GreptimeDB;
+import io.greptime.metric.Memory;
+import io.greptime.StreamWriter;
+import io.greptime.WriteOp;
 import io.greptime.models.WriteOk;
+import io.greptime.quickstart.TestConnector;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
