+++ ingester-example/src/main/java/io/greptime/quickstart/write/LowLevelApiStreamWriteQuickStart.java
@@ -14,12 +14,16 @@
  * limitations under the License.
  */
 
-package io.greptime;
+package io.greptime.quickstart.write;
 
+import io.greptime.GreptimeDB;
+import io.greptime.StreamWriter;
+import io.greptime.WriteOp;
 import io.greptime.models.DataType;
 import io.greptime.models.Table;
 import io.greptime.models.TableSchema;
 import io.greptime.models.WriteOk;
+import io.greptime.quickstart.TestConnector;
 import io.greptime.rpc.Compression;
 import io.greptime.rpc.Context;
 import java.util.concurrent.CompletableFuture;
