+++ ingester-example/src/main/java/io/greptime/quickstart/write/LowLevelApiWriteQuickStart.java
@@ -14,12 +14,17 @@
  * limitations under the License.
  */
 
-package io.greptime;
+package io.greptime.quickstart.write;
 
+import io.greptime.GreptimeDB;
+import io.greptime.WriteOp;
+import io.greptime.models.DataType;
 import io.greptime.models.Err;
 import io.greptime.models.Result;
+import io.greptime.models.Table;
+import io.greptime.models.TableSchema;
 import io.greptime.models.WriteOk;
-import java.util.ArrayList;
+import io.greptime.quickstart.TestConnector;
 import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
@@ -28,45 +33,72 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * This example demonstrates how to use the high-level API to write data to the database.
+ * This example demonstrates how to use the low-level API to write data to the database.
  * It shows how to define the schema for metrics tables, write data to the table, and get the write result.
  * It also shows how to delete data from the table using the `WriteOp.Delete`.
  */
-public class HighLevelApiWriteQuickStart {
+public class LowLevelApiWriteQuickStart {
 
-    private static final Logger LOG = LoggerFactory.getLogger(HighLevelApiWriteQuickStart.class);
+    private static final Logger LOG = LoggerFactory.getLogger(LowLevelApiWriteQuickStart.class);
 
     public static void main(String[] args) throws ExecutionException, InterruptedException {
         GreptimeDB greptimeDB = TestConnector.connectToDefaultDB();
 
-        List<Cpu> cpus = new ArrayList<>();
+        // Define the schema for metrics tables.
+        // The schema is immutable and can be safely reused across multiple operations.
+        // It is recommended to use snake_case for column names.
+        TableSchema cpuMetricSchema = TableSchema.newBuilder("cpu_metric")
+                .addTag("host", DataType.String)
+                .addTimestamp("ts", DataType.TimestampMillisecond)
+                .addField("cpu_user", DataType.Float64)
+                .addField("cpu_sys", DataType.Float64)
+                .build();
+
+        TableSchema memMetricSchema = TableSchema.newBuilder("mem_metric")
+                .addTag("host", DataType.String)
+                .addTimestamp("ts", DataType.TimestampMillisecond)
+                .addField("mem_usage", DataType.Float64)
+                .build();
+
+        // Tables are not reusable - a new instance must be created for each write operation.
+        // However, we can add multiple rows to a single table before writing it,
+        // which is more efficient than writing rows individually.
+        Table cpuMetric = Table.from(cpuMetricSchema);
+        Table memMetric = Table.from(memMetricSchema);
+
         for (int i = 0; i < 10; i++) {
-            Cpu c = new Cpu();
-            c.setHost("127.0.0." + i);
-            c.setTs(System.currentTimeMillis());
-            c.setCpuUser(i + 0.1);
-            c.setCpuSys(i + 0.12);
-            cpus.add(c);
+            String host = "127.0.0." + i;
+            long ts = System.currentTimeMillis();
+            double cpuUser = i + 0.1;
+            double cpuSys = i + 0.12;
+            // Add a row to the `cpu_metric` table.
+            // The order of the values must match the schema definition.
+            cpuMetric.addRow(host, ts, cpuUser, cpuSys);
         }
 
-        List<Memory> memories = new ArrayList<>();
         for (int i = 0; i < 10; i++) {
-            Memory m = new Memory();
-            m.setHost("127.0.0." + i);
-            m.setTs(System.currentTimeMillis());
-            m.setMemUsage(i + 0.2);
-            memories.add(m);
+            String host = "127.0.0." + i;
+            long ts = System.currentTimeMillis();
+            double memUsage = i + 0.2;
+            // Add a row to the `mem_metric` table.
+            // The order of the values must match the schema definition.
+            memMetric.addRow(host, ts, memUsage);
         }
+
+        // Complete the table to make it immutable. If users forget to call this method,
+        // it will still be called internally before the table data is written.
+        cpuMetric.complete();
+        memMetric.complete();
 
         // For performance reasons, the SDK is designed to be purely asynchronous.
         // The return value is a CompletableFuture object. If you want to immediately obtain
         // the result, you can call `future.get()`, which will block until the operation completes.
         // For production environments, consider using non-blocking approaches with callbacks or
         // the CompletableFuture API.
-        CompletableFuture<Result<WriteOk, Err>> puts = greptimeDB.writeObjects(cpus, memories);
+        CompletableFuture<Result<WriteOk, Err>> future = greptimeDB.write(cpuMetric, memMetric);
 
         // Now we can get the write result.
-        Result<WriteOk, Err> result = puts.get();
+        Result<WriteOk, Err> result = future.get();
 
         // The Result object holds either a success value (WriteOk) or an error (Err).
         // We can transform these values using the `map` method for success cases and `mapErr` for error cases.
@@ -80,10 +112,10 @@ public class HighLevelApiWriteQuickStart {
             LOG.error("Failed to write: {}", simpleResult.getErr());
         }
 
-        List<List<?>> deletePojoObjects = Arrays.asList(cpus.subList(0, 5), memories.subList(0, 5));
+        List<Table> delete_objs = Arrays.asList(cpuMetric.subRange(0, 5), memMetric.subRange(0, 5));
         // We can also delete data from the table using the `WriteOp.Delete`.
         Result<WriteOk, Err> deletes =
-                greptimeDB.writeObjects(deletePojoObjects, WriteOp.Delete).get();
+                greptimeDB.write(delete_objs, WriteOp.Delete).get();
 
         if (deletes.isOk()) {
             LOG.info("Delete result: {}", result.getOk());
