+++ ingester-example/src/main/resources/db-connection.properties
@@ -15,5 +15,7 @@
 # DataSource
 db.database-driver=com.mysql.cj.jdbc.Driver
 db.url=jdbc:mysql://localhost:4002/public
+db.endpoints=localhost:4001
+db.database=public
 db.username=
 db.password=
